re-org urn logic
Move validation to urn.ts.
And remove shorten function.
